PreSetService: PreSetController 내용 추가 fix: HabitDtoImpl 에 List<Integer>, Category enum타입 추가

+++ src/main/java/com/sollertia/habit/domain/preset/dto/PreSetVo.java
@@ -15,7 +15,7 @@ public class PreSetVo {
     private Long categoryId;
     private String title;
     private String description;
-    private Long period;
+    private int period;
     private Long count;
     private Category category;
     private String session;
